+++ packages/perplexity/src/perplexity-language-model.ts
@@ -148,13 +148,17 @@ export class PerplexityLanguageModel implements LanguageModelV1 {
   ): Promise<Awaited<ReturnType<LanguageModelV1['doGenerate']>>> {
     const { args, warnings } = this.getArgs(options);
 
-    const { responseHeaders, value: response } = await postJsonToApi({
+    const {
+      responseHeaders,
+      value: response,
+      rawValue: rawResponse,
+    } = await postJsonToApi({
       url: `${this.config.baseURL}/chat/completions`,
       headers: combineHeaders(this.config.headers(), options.headers),
       body: args,
       failedResponseHandler: createJsonErrorResponseHandler({
         errorSchema: perplexityErrorSchema,
-        errorToMessage: data => data.error,
+        errorToMessage,
       }),
       successfulResponseHandler: createJsonResponseHandler(
         perplexityResponseSchema,
@@ -176,7 +180,7 @@ export class PerplexityLanguageModel implements LanguageModelV1 {
         completionTokens: response.usage.completion_tokens,
       },
       rawCall: { rawPrompt, rawSettings },
-      rawResponse: { headers: responseHeaders },
+      rawResponse: { headers: responseHeaders, body: rawResponse },
       request: { body: JSON.stringify(args) },
       response: getResponseMetadata(response),
       warnings,
@@ -216,7 +220,7 @@ export class PerplexityLanguageModel implements LanguageModelV1 {
       body,
       failedResponseHandler: createJsonErrorResponseHandler({
         errorSchema: perplexityErrorSchema,
-        errorToMessage: data => data.error,
+        errorToMessage,
       }),
       successfulResponseHandler: createEventSourceResponseHandler(
         perplexityChunkSchema,
@@ -422,8 +426,15 @@ const perplexityChunkSchema = z.object({
 });
 
 export const perplexityErrorSchema = z.object({
-  code: z.string(),
-  error: z.string(),
+  error: z.object({
+    code: z.number(),
+    message: z.string().nullish(),
+    type: z.string().nullish(),
+  }),
 });
 
 export type PerplexityErrorData = z.infer<typeof perplexityErrorSchema>;
+
+const errorToMessage = (data: PerplexityErrorData) => {
+  return data.error.message ?? data.error.type ?? 'unknown error';
+};