+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,27 @@
 
 ### Patch Changes
 
+- Updated dependencies [ee1c787]
+  - ai@4.1.54
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [e1d3d42]
+  - ai@4.1.53
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5329a69]
+  - ai@4.1.52
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [0cb2647]
   - ai@4.1.51