+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/valibot
 
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [ee1c787]
+  - ai@4.1.54
+
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [e1d3d42]
+  - ai@4.1.53
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [5329a69]
+  - ai@4.1.52
+
 ## 0.0.18
 
 ### Patch Changes